consolidate search tests
